test: restore fstab/crypttab for RHEL 8 storage tests
When re-trying non-destructive tests the testFstabOptions test fails as
/etc/fstab is not restored to its pristine state. CoreOS does not run
storage tests so should not restore fstab/crypttab and the ws-container
scenario should not stop cockpit as it isn't installed on the host
itself.

+++ test/common/testlib.py
@@ -1816,13 +1816,14 @@ class MachineCase(unittest.TestCase):
         # cockpit configuration
         self.restore_dir("/etc/cockpit")
 
-        if not m.ws_container:
+        if not m.ostree_image:
             # for storage tests
             self.restore_file("/etc/fstab")
             self.restore_file("/etc/crypttab")
 
-            # tests expect cockpit.service to not run at start; also, avoid log leakage into the next test
-            self.addCleanup(m.execute, "systemctl stop --quiet cockpit")
+            if not m.ws_container:
+                # tests expect cockpit.service to not run at start; also, avoid log leakage into the next test
+                self.addCleanup(m.execute, "systemctl stop --quiet cockpit")
 
         # The sssd daemon seems to get confused when we restore
         # backups of /etc/group etc and stops following updates to it.
